rename history, resolve TxHistory collision
